Attempts at simple optimization.

+++ ChessEngineLibrary/moveTree.h
@@ -21,4 +21,4 @@ struct searchData
 
 Move startSearch(int searchDepth, Board board, colours colour);
 Move rootSearch(int depthLeft, Board board, colours colour);
-int negamax(int alpha, int beta, int depthLeft, Board board, colours colour, searchData* data, int hash, bool isQuiet);
+int negamax(int alpha, int beta, int depthLeft, Board board, colours colour, searchData* data, uint64_t hash, bool isQuiet);
